Minor fixes

+++ Elastic/GPU/swapbytes.h
@@ -6,9 +6,9 @@ extern "C"
 {
 #endif
 
-void swap2bytes(short *i2, int n);
-void swap4bytes(int *i4, int n);
-void swap8bytes(long* i8, int n);
+void swap2bytes(short *i2, long n);
+void swap4bytes(int *i4, long n);
+void swap8bytes(long* i8, long n);
 
 #ifdef __cplusplus
 }
